requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
pymongo==4.6.1
dnspython==2.4.2
python-dotenv==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
authlib==1.3.0
requests==2.31.0
